Moving everything over to a diff dir again
