+++ nextjs-blog-starter-mdx/src/design/avatar.js
