+++ nextjs-blog-starter-mdx/src/design/cover-image.js
