+++ nextjs-blog-starter-mdx/src/design/footer.js
