+++ nextjs-blog-starter-mdx/src/design/header.js
