+++ nextjs-blog-starter-mdx/src/design/intro.js
