+++ nextjs-blog-starter-mdx/src/design/layout.js
