+++ nextjs-blog-starter-mdx/src/design/more-stories.js
