+++ nextjs-blog-starter-mdx/src/design/post-body.js
